images saved in the database

+++ src/pages/MerchantPage/MerchantHomePage/context/AddhotelContext.jsx
@@ -14,6 +14,7 @@ export const AddHotelProvider = ({ children }) => {
   const [editHotelState, setEditHotelState] = useState({});
   const [editRoomState, setEditRoomState] = useState({});
   const [getHotelId, setGetHotelId] = useState();
+
   const [fetchUserListing, setFetchUserListing] = useState({
     fetching: true,
     responseData: [],
@@ -45,6 +46,7 @@ export const AddHotelProvider = ({ children }) => {
     price: 0,
     maxPeople: 0,
     roomNumbers: [{ number: 0 }],
+    photos: [],
   });
   //   const { loading, setLoading } = useState(false);
   const toggleModal = () => {
@@ -92,47 +94,103 @@ export const AddHotelProvider = ({ children }) => {
   const submitListings = async () => {
     let url = `${WILL_TRIP_BASE_URL}/hotels/merchant`;
 
-    let response = await axios.post(
-      url,
-      {
-        ...listingsData,
-        userID: user?.id,
-      },
-      {
+    const formData = new FormData();
+    formData.append("name", listingsData.name);
+    formData.append("destination", listingsData.destination);
+    formData.append("feature", listingsData.feature);
+    formData.append("address", listingsData.address);
+    formData.append("price", listingsData.price);
+    formData.append("overview", listingsData.overview);
+    formData.append("facilities", listingsData.facilities);
+    formData.append("foods_and_drinks", listingsData.foods_and_drinks);
+    formData.append("location", listingsData.location);
+
+    for (let i = 0; i < listingsData.photos.length; i++) {
+      formData.append("photos", listingsData.photos[i]);
+    }
+    formData.append("userID", user?.id);
+    try {
+      const response = await axios.post(url, formData, {
         headers: {
           Authorization: `Bearer ${user?.token}`,
+          "Content-Type": "multipart/form-data",
         },
+      });
+      if (response?.data?.status === "success") {
+        toast.success(response?.data?.msg);
+        toggleModal();
+        initializeState();
+      } else {
+        toggleModal();
       }
-    );
-    if (response?.data?.status === "success") {
-      toast.success(response?.data?.msg);
-      toggleModal();
-      initializeState();
-    } else {
+    } catch (error) {
+      console.log(error);
       toggleModal();
     }
   };
 
-  const addRooms = async () => {
-    let url = `${WILL_TRIP_BASE_URL}/rooms/${getHotelId}/merchant`;
+  // const addRooms = async () => {
+  //   let url = `${WILL_TRIP_BASE_URL}/rooms/${getHotelId}/merchant`;
 
-    let response = await axios.post(
-      url,
-      {
-        ...roomsListings,
-        userID: user?.id,
-      },
-      {
+  //   let response = await axios.post(
+  //     url,
+  //     {
+  //       ...roomsListings,
+  //       userID: user?.id,
+  //     },
+  //     {
+  //       headers: {
+  //         Authorization: `Bearer ${user?.token}`,
+  //         "Content-Type": "multipart/form-data",
+  //       },
+  //     }
+  //   );
+  //   if (response?.data?.status === "success") {
+  //     toast.success(response?.data?.msg);
+  //     addRoomsModal();
+  //     initializeState();
+  //   } else {
+  //     addRoomsModal();
+  //   }
+  // };
+
+  const addRooms = async () => {
+    const url = `${WILL_TRIP_BASE_URL}/rooms/${getHotelId}/merchant`;
+
+    const formData = new FormData();
+    formData.append("title", roomsListings.title);
+    formData.append("price", roomsListings.price);
+    formData.append("description", roomsListings.description);
+    formData.append("maxPeople", roomsListings.maxPeople);
+    for (let i = 0; i < roomsListings.roomNumbers.length; i++) {
+      formData.append(
+        `roomNumbers[${i}][number]`,
+        roomsListings.roomNumbers[i].number
+      );
+    }
+
+    for (let i = 0; i < roomsListings.photos.length; i++) {
+      formData.append("photos", roomsListings.photos[i]);
+    }
+    formData.append("userID", user?.id);
+
+    try {
+      const response = await axios.post(url, formData, {
         headers: {
           Authorization: `Bearer ${user?.token}`,
+          "Content-Type": "multipart/form-data",
         },
+      });
+
+      if (response?.data?.status === "success") {
+        toast.success(response?.data?.msg);
+        addRoomsModal();
+        initializeState();
+      } else {
+        addRoomsModal();
       }
-    );
-    if (response?.data?.status === "success") {
-      toast.success(response?.data?.msg);
-      addRoomsModal();
-      initializeState();
-    } else {
+    } catch (error) {
+      console.log(error);
       addRoomsModal();
     }
   };
@@ -165,18 +223,20 @@ export const AddHotelProvider = ({ children }) => {
 
   const deleteHotel = async (id) => {
     let url = `${WILL_TRIP_BASE_URL}/hotels/merchant/${id}`;
-    try {
-      let response = await axios.delete(url, {
-        headers: {
-          Authorization: `Bearer ${user?.token}`,
-        },
-      });
-      if (response.data.status === "success") {
-        initializeState();
-        toast.success(response?.data?.msg);
+    if (window.confirm("Are you sure you want to proceed?")) {
+      try {
+        let response = await axios.delete(url, {
+          headers: {
+            Authorization: `Bearer ${user?.token}`,
+          },
+        });
+        if (response.data.status === "success") {
+          initializeState();
+          toast.success(response?.data?.msg);
+        }
+      } catch (error) {
+        toast.error(error?.response?.data?.message);
       }
-    } catch (error) {
-      toast.error(error?.response?.data?.message);
     }
   };
 
@@ -187,6 +247,7 @@ export const AddHotelProvider = ({ children }) => {
         photos: e.target.files,
       }));
     }
+
     if (!e.target.files) {
       setListingsData((prev) => ({
         ...prev,
@@ -196,10 +257,19 @@ export const AddHotelProvider = ({ children }) => {
   };
 
   const addRoomsOnChange = (e) => {
-    setRoomsListings((prev) => ({
-      ...prev,
-      [e.target.id]: e.target.value,
-    }));
+    if (e.target.files) {
+      setRoomsListings((prev) => ({
+        ...prev,
+        photos: e.target.files,
+      }));
+    }
+
+    if (!e.target.files) {
+      setRoomsListings((prev) => ({
+        ...prev,
+        [e.target.id]: e.target.value,
+      }));
+    }
   };
 
   let initializeState = () => {
